fix: move migrations and indices
Moved postgres migrations and elastic indices into separate folder outside of initializer
package to be accessible from everywhere.

A lot of things went wrong during the process. Lot of typescript cofig had to be changed.
