+++ docker-compose.dev.yaml
@@ -23,6 +23,11 @@ services:
       - esdata:/usr/share/elasticsearch/data
 
   initializer:
+    build:
+      target: dev
+    volumes:
+      - ./packages/initializer/src:/app/packages/initializer/src:z
+      - ./packages/shared:/app/packages/shared:z
 
   kibana:
