+++ docker-compose.prod.yaml
@@ -19,7 +19,8 @@ services:
       - esdata:/usr/share/elasticsearch/data
 
   initializer:
-
+    build:
+      target: runtime
   kibana:
 
   postgres:
