+++ package.json
@@ -42,6 +42,7 @@
   },
   "private": true,
   "dependencies": {
+    "@vera/shared": "workspace:*",
     "ajv": "^8.18.0",
     "vue": "^3.5.28"
   }
