+++ packages/shared/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@vera/shared",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "packageManager": "pnpm@10.11.0",
+  "dependencies": {
+    "@elastic/elasticsearch": "^9.3.1",
+    "kysely": "^0.28.11"
+  },
+  "type": "module"
+}
