+++ packages/shared/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "noEmit": true,
+    "composite": true,
+    "declaration": true,
+    "rootDir": ".",
+  },
+}
